microsservices/catalog/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-dotenv
pytest
httpx
pytest-asyncio
